Change: VMCMixer.Backend.mainLoop> Terminate all 'awaitPacket' asyncs when it receives async's cancellation
In previous implementation, I couldn't access and terminate
those asyncs when 'mainLoop' is cancelled because 'mainLoop' can't
return it's output when itself is cancelled.

As async's cancellation is done by throwing asyncrhonized exception,
I can 'catch' it inside and do whatever I have to do as cleanup
operation.

Although diff is a bit big, not much things are changed.

Here is a rough summary of changes:
1. 'go' function will detect 'AsyncCancelled' exception and return children 'Async's that need to be cancelled.
2. 'mainLoop' itself will receive those 'Async's and 'cancel' all of them before finishing its operation.